Move the SavedStateFlow code to the FlowExt.kt
