+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/ui/orders/details/OrderDetailsViewModel.kt
@@ -3,6 +3,7 @@ package com.woocommerce.android.ui.orders.details
 import android.os.Parcelable
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.asLiveData
+import com.woocommerce.android.extensions.getStateFlow
 import com.woocommerce.android.ui.NavArgs.ORDER_ID
 import com.woocommerce.android.ui.login.LoginRepository
 import com.woocommerce.android.ui.orders.FormatOrderData
@@ -11,7 +12,6 @@ import com.woocommerce.android.ui.orders.OrdersRepository
 import com.woocommerce.android.ui.orders.details.FetchOrderProducts.OrderProductsRequest.Error
 import com.woocommerce.android.ui.orders.details.FetchOrderProducts.OrderProductsRequest.Finished
 import com.woocommerce.android.viewmodel.ScopedViewModel
-import com.woocommerce.commons.viewmodel.getStateFlow
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.launchIn
